+++ src/context.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const WaitingContext = createContext();
+
+export default WaitingContext;
